requirements.txt
fastapi==0.114.2
timm==1.0.9
torch==2.4.1
torchvision==0.19.1
pillow==10.4.0
requests==2.32.3
qdrant_client==1.11.2
sentence_transformers==3.1.0
uvicorn==0.30.6